abstractNumberModel를 생성하여 스켈레톤 골격 구현
